Diagnosis script updated address PEDSnet CDM2.6
fixes #256

PEDSnet CDM2.6
guidance on populating the condition_concept_id, condition_source_concept_id, condition_source_value. These change was present in the change documentation for v2.3 to v2.4 but not as explicitly defined. The conventions below were made compulsory for version 2.6.

Diagnosis Name "|" IMO Code "|" Diagnosis Code

This changes the query in the script. So following script was modified.