Administrative visit added to concept map
fixes #
Updated and added the `Administrative visit` in concept map as `OT` in `Discharge disposition`
changes are made as per PEDSnet CDM2.6 as below

`Addition of "Administrative Visit" Visit Type. This is a new concept id option for the visit_concept_id field. Please see NOTE 1 for ETL guidance on the various visit categories.`